Don't add remaining contents to closed containers (fixes #57)
The submitted test case failed because the parser tried to add an empty line to the already-closed IndentedCode block.
In the JS reference parser, closing this block type results in it becoming a CodeBlock type, which we don't use, and therefore
the addLine() logic doesn't run.  We can mirror that behavior by ensuring nothing ever calls addLine() on a closed block.