Fix XSS vulnerability caused by improper preservation of entities (#353)
The upstream library we forked from had a bug where entities were
incorrectly preserved when rendering href, src, title, and info string
attributes.

This gave rise to double-encoding errors, when the original markdown is
e.g. `&amp;colon;`, since the parsers already unescape entities.

Thanks to Sebastiaan Knijnenburg for reporting this and for John
MacFarlane for providing the fix.