Improve DefaultSlugGenerator to preserve complex unicode letters, numbers, and marks (#467)
When the slug is generated from the title text, try to preserve letters, numbers and marks declared in Unicode spec.

Certain complex languages declare code points for marks that are supposed to be used with letters. This allows to use titles in every Unicode-recognized language to safely generate a slug from the title text without dissecting to a point that it loses the meaning.