Include composer autoloader in example code
If someone is new with composer (which is entirely possible) it might
be fconfusing to get a "class not found" error when copy/pasting the
example code.

This commit adds including the autoloader to the examples for basic
usage, so they can be literally copy/pasted in a text editor and will
work right away.

This can make it easier to get started for beginners.