Add instructions on how to install dev version
When version 2.0 goes stable, it will be as easy as is currently
documented:

```bash
composer require "league/commonmark:^2.0"
```

Right now however, there's a good chance people have their composer
settings set to have a minimum stability of "stable".
This will probably result in errors when trying to run this command.

I wasn't sure what the best way forward is for here: leave in the old,
but make a note of it? Make temporary instructions for dev?

For now I went with a combination: leave the current way (with just
`composer require`), but add some instructions for the current dev
version.

When this version goes stable the extra instructions can easily be
deleted from the document.